+++ lib/app_routes/route_path.dart
@@ -1,7 +1,7 @@
 class RoutePath {
   RoutePath._();
 
-  static const String productDetailsView = '/products';
-
   static const String homeView = '/';
+  static const String productDetailsView = '/product_details_view';
+  static const String userDetailsView = '/user_details_view';
 }
