+++ lib/main.dart
@@ -2,33 +2,14 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import 'app_routes/router.dart';
+import 'my_app.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return ProviderScope(
-      child: MaterialApp.router(
-        title: 'Riverpod',
-        debugShowCheckedModeBanner: false,
-        theme: ThemeData(
-          colorScheme: ColorScheme.fromSeed(seedColor: Colors.blue),
-          useMaterial3: true,
-        ),
-        routerDelegate: router.routerDelegate,
-        routeInformationParser: router.routeInformationParser,
-        backButtonDispatcher: router.backButtonDispatcher,
-        routeInformationProvider: router.routeInformationProvider,
-      ),
-    );
-  }
+  runApp(
+    const ProviderScope(
+      child: MyApp(),
+    ),
+  );
 }
