+++ pubspec.yaml
@@ -43,6 +43,7 @@ dependencies:
   google_fonts: ^6.2.1
   snackbar: ^1.0.2
   awesome_snackbar_content: ^0.1.4
+  firebase_core: ^3.9.0
 
 
 dev_dependencies:
